refactor: Seggregates MasteryFactory from KcProgress service.
